+++ src/김근범/백트래킹/week1/금요일/도영이가_만든_맛있는_음식.java
@@ -1,4 +1,4 @@
-package 김근범.백트래킹.금요일;
+package 김근범.백트래킹.week1.금요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
