+++ src/김근범/백트래킹/week1/금요일/카드놓기.java
@@ -1,4 +1,4 @@
-package 김근범.백트래킹.금요일;
+package 김근범.백트래킹.week1.금요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
