+++ src/김근범/백트래킹/week1/목요일/N과M_11.java
@@ -1,4 +1,4 @@
-package 김근범.백트래킹.화요일;
+package 김근범.백트래킹.week1.목요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,45 +6,44 @@ import java.io.InputStreamReader;
 import java.util.Arrays;
 import java.util.StringTokenizer;
 
-public class N과M_5 {
+public class N과M_11 {
 
     static int N, M;
     static int[] arr;
     static int[] result;
-    static boolean[] check;
     static StringBuilder sb = new StringBuilder();
-
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
         N = Integer.parseInt(st.nextToken());
         M = Integer.parseInt(st.nextToken());
-        st = new StringTokenizer(br.readLine());
         arr = new int[N];
-        check = new boolean[N];
         result = new int[M];
+        st = new StringTokenizer(br.readLine());
         for(int i=0;i<N;i++){
             arr[i] = Integer.parseInt(st.nextToken());
         }
         Arrays.sort(arr);
-        fetchSequence(0);
+        fetchSequences(0);
         System.out.print(sb);
     }
 
-    private static void fetchSequence(int index) {
+    private static void fetchSequences(int index) {
         if(index == M) {
-            for(int i : result) {
+            for(int i: result) {
                 sb.append(i).append(' ');
             }
             sb.append('\n');
             return;
         }
-        for(int i = 0; i<N;i++){
-            if(!check[i]) {
-                check[i] = true;
+
+        int prev = -1;
+
+        for(int i=0;i<N;i++){
+            if(prev!=arr[i]){
+                prev = arr[i];
                 result[index] = arr[i];
-                fetchSequence(index+1);
-                check[i] = false;
+                fetchSequences(index+1);
             }
         }
     }
