+++ src/김근범/백트래킹/week1/목요일/N과M_12.java
@@ -1,4 +1,4 @@
-package 김근범.백트래킹.화요일;
+package 김근범.백트래킹.week1.목요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,13 +6,13 @@ import java.io.InputStreamReader;
 import java.util.Arrays;
 import java.util.StringTokenizer;
 
-public class N과M_8 {
+public class N과M_12 {
 
     static int N, M;
     static int[] arr;
     static int[] result;
-
     static StringBuilder sb = new StringBuilder();
+
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
@@ -21,28 +21,30 @@ public class N과M_8 {
         arr = new int[N];
         result = new int[M];
         st = new StringTokenizer(br.readLine());
-        for(int i=0;i<N;i++){
+        for (int i = 0; i < N; i++) {
             arr[i] = Integer.parseInt(st.nextToken());
         }
         Arrays.sort(arr);
-        fetchSequence(0);
-        System.out.println(sb);
+        fetchSequences(0);
+        System.out.print(sb);
     }
 
-    private static void fetchSequence(int index) {
-        if(index == M) {
-            for(int i: result) {
+    private static void fetchSequences(int index) {
+        if (index == M) {
+            for (int i : result) {
                 sb.append(i).append(' ');
             }
             sb.append('\n');
             return;
         }
 
+        int prev = -1;
 
-        for(int i=0;i<N;i++) {
-            if(index ==0 || result[index-1] <= arr[i]) {
+        for (int i = 0; i < N; i++) {
+            if(prev != arr[i] && (index==0 || result[index-1] <= arr[i])) {
+                prev = arr[i];
                 result[index] = arr[i];
-                fetchSequence(index+1);
+                fetchSequences(index+1);
             }
         }
     }
