+++ src/김근범/백트래킹/week1/수요일/N과M_6.java
@@ -1,16 +1,15 @@
-package 김근범.백트래킹.수요일;
+package 김근범.백트래킹.week1.수요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.*;
+import java.util.Arrays;
+import java.util.StringTokenizer;
 
-public class N과M_9 {
+public class N과M_6 {
     static int N, M;
-    static LinkedHashSet<Integer> set = new LinkedHashSet<>();
     static int[] arr;
     static int[] result;
-    static HashMap<Integer, Integer> map = new HashMap<>();
     static StringBuilder sb = new StringBuilder();
 
     public static void main(String[] args) throws IOException {
@@ -18,17 +17,13 @@ public class N과M_9 {
         StringTokenizer st = new StringTokenizer(br.readLine());
         N = Integer.parseInt(st.nextToken());
         M = Integer.parseInt(st.nextToken());
-        result = new int[M];
         arr = new int[N];
+        result = new int[M];
         st = new StringTokenizer(br.readLine());
         for(int i=0;i<N;i++){
             arr[i] = Integer.parseInt(st.nextToken());
         }
         Arrays.sort(arr);
-        for(int i=0;i<N;i++){
-            set.add(arr[i]);
-            map.put(arr[i], map.getOrDefault(arr[i], 0) + 1);
-        }
         fetchSequences(0);
         System.out.print(sb);
     }
@@ -42,12 +37,10 @@ public class N과M_9 {
             return;
         }
 
-        for (int num : set) {
-            if (map.getOrDefault(num, 0) > 0) {
-                map.put(num, map.get(num) - 1);
-                result[index] = num;
-                fetchSequences(index + 1);
-                map.put(num, map.get(num) + 1);
+        for(int i=0;i<N;i++){
+            if(index == 0 || result[index-1] < arr[i]) {
+                result[index] = arr[i];
+                fetchSequences(index+1);
             }
         }
     }
