+++ src/김근범/백트래킹/week1/수요일/N과M_7.java
@@ -1,4 +1,4 @@
-package 김근범.백트래킹.목요일;
+package 김근범.백트래킹.week1.수요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,8 +6,7 @@ import java.io.InputStreamReader;
 import java.util.Arrays;
 import java.util.StringTokenizer;
 
-public class N과M_12 {
-
+public class N과M_7 {
     static int N, M;
     static int[] arr;
     static int[] result;
@@ -21,7 +20,7 @@ public class N과M_12 {
         arr = new int[N];
         result = new int[M];
         st = new StringTokenizer(br.readLine());
-        for (int i = 0; i < N; i++) {
+        for(int i=0;i<N;i++){
             arr[i] = Integer.parseInt(st.nextToken());
         }
         Arrays.sort(arr);
@@ -30,22 +29,17 @@ public class N과M_12 {
     }
 
     private static void fetchSequences(int index) {
-        if (index == M) {
-            for (int i : result) {
+        if(index == M) {
+            for(int i : result) {
                 sb.append(i).append(' ');
             }
             sb.append('\n');
             return;
         }
 
-        int prev = -1;
-
-        for (int i = 0; i < N; i++) {
-            if(prev != arr[i] && (index==0 || result[index-1] <= arr[i])) {
-                prev = arr[i];
-                result[index] = arr[i];
-                fetchSequences(index+1);
-            }
+        for(int i=0;i<N;i++){
+            result[index] = arr[i];
+            fetchSequences(index+1);
         }
     }
 }
