+++ src/김근범/백트래킹/week1/월요일/N과M_3.java
@@ -1,4 +1,4 @@
-package 김근범.백트래킹.월요일;
+package 김근범.백트래킹.week1.월요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
