+++ src/김근범/백트래킹/week1/월요일/N과M_4.java
@@ -1,48 +1,36 @@
-package 김근범.백트래킹.목요일;
+package 김근범.백트래킹.week1.월요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.Arrays;
 import java.util.StringTokenizer;
 
-public class N과M_11 {
-
+public class N과M_4 {
     static int N, M;
-    static int[] arr;
-    static int[] result;
+    static int[] sequence;
     static StringBuilder sb = new StringBuilder();
+
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        StringTokenizer st = new StringTokenizer(br.readLine());
+        StringTokenizer st =new StringTokenizer(br.readLine());
         N = Integer.parseInt(st.nextToken());
         M = Integer.parseInt(st.nextToken());
-        arr = new int[N];
-        result = new int[M];
-        st = new StringTokenizer(br.readLine());
-        for(int i=0;i<N;i++){
-            arr[i] = Integer.parseInt(st.nextToken());
-        }
-        Arrays.sort(arr);
+        sequence = new int[M];
         fetchSequences(0);
         System.out.print(sb);
     }
 
     private static void fetchSequences(int index) {
         if(index == M) {
-            for(int i: result) {
+            for (int i : sequence) {
                 sb.append(i).append(' ');
             }
             sb.append('\n');
             return;
         }
-
-        int prev = -1;
-
-        for(int i=0;i<N;i++){
-            if(prev!=arr[i]){
-                prev = arr[i];
-                result[index] = arr[i];
+        for(int i=1;i<=N;i++){
+            if(index==0 || sequence[index-1] <= i) {
+                sequence[index] = i;
                 fetchSequences(index+1);
             }
         }
