+++ src/김근범/백트래킹/week1/월요일/연산자끼워넣기.java
@@ -1,4 +1,4 @@
-package 김근범.백트래킹.월요일;
+package 김근범.백트래킹.week1.월요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
