+++ src/김근범/백트래킹/week1/화요일/N과M_5.java
@@ -1,4 +1,4 @@
-package 김근범.백트래킹.목요일;
+package 김근범.백트래킹.week1.화요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,7 +6,8 @@ import java.io.InputStreamReader;
 import java.util.Arrays;
 import java.util.StringTokenizer;
 
-public class N과M_10 {
+public class N과M_5 {
+
     static int N, M;
     static int[] arr;
     static int[] result;
@@ -18,19 +19,19 @@ public class N과M_10 {
         StringTokenizer st = new StringTokenizer(br.readLine());
         N = Integer.parseInt(st.nextToken());
         M = Integer.parseInt(st.nextToken());
-        arr = new int[N];
-        result = new int[M];
-        check = new boolean[N];
         st = new StringTokenizer(br.readLine());
-        for (int i = 0; i < N; i++) {
+        arr = new int[N];
+        check = new boolean[N];
+        result = new int[M];
+        for(int i=0;i<N;i++){
             arr[i] = Integer.parseInt(st.nextToken());
         }
         Arrays.sort(arr);
-        fetchSequences(0);
+        fetchSequence(0);
         System.out.print(sb);
     }
 
-    private static void fetchSequences(int index) {
+    private static void fetchSequence(int index) {
         if(index == M) {
             for(int i : result) {
                 sb.append(i).append(' ');
@@ -38,15 +39,11 @@ public class N과M_10 {
             sb.append('\n');
             return;
         }
-
-        int prev = -1;
-
-        for(int i=0;i<N;i++){
-            if((index == 0 || (!check[i] && result[index-1] <= arr[i])) && prev != arr[i]) {
-                result[index] = arr[i];
-                prev = arr[i];
+        for(int i = 0; i<N;i++){
+            if(!check[i]) {
                 check[i] = true;
-                fetchSequences(index+1);
+                result[index] = arr[i];
+                fetchSequence(index+1);
                 check[i] = false;
             }
         }
