+++ src/김근범/백트래킹/week1/화요일/N과M_8.java
@@ -1,4 +1,4 @@
-package 김근범.백트래킹.수요일;
+package 김근범.백트래킹.week1.화요일;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,12 +6,13 @@ import java.io.InputStreamReader;
 import java.util.Arrays;
 import java.util.StringTokenizer;
 
-public class N과M_7 {
+public class N과M_8 {
+
     static int N, M;
     static int[] arr;
     static int[] result;
-    static StringBuilder sb = new StringBuilder();
 
+    static StringBuilder sb = new StringBuilder();
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
@@ -24,22 +25,25 @@ public class N과M_7 {
             arr[i] = Integer.parseInt(st.nextToken());
         }
         Arrays.sort(arr);
-        fetchSequences(0);
-        System.out.print(sb);
+        fetchSequence(0);
+        System.out.println(sb);
     }
 
-    private static void fetchSequences(int index) {
+    private static void fetchSequence(int index) {
         if(index == M) {
-            for(int i : result) {
+            for(int i: result) {
                 sb.append(i).append(' ');
             }
             sb.append('\n');
             return;
         }
 
-        for(int i=0;i<N;i++){
-            result[index] = arr[i];
-            fetchSequences(index+1);
+
+        for(int i=0;i<N;i++) {
+            if(index ==0 || result[index-1] <= arr[i]) {
+                result[index] = arr[i];
+                fetchSequence(index+1);
+            }
         }
     }
 }
